+++ src/backward-compatibility.js
