+++ src/conversions/ascii.js
