+++ src/conversions/data/text-operators.json
