+++ src/conversions/svg.js
