+++ src/translations/nn.json
