+++ src/translations/no.json
