+++ tests/guidelines/2024-1/item1-namespace.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 1: Namespace and Semantics Handling', () => {
   beforeEach(() => {
