+++ tests/guidelines/2024-1/item11-units-and-numbers.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 11: Units and Numbers', () => {
   describe('Unit Handling', () => {
