+++ tests/guidelines/2024-1/item2-mfenced.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 2: mfenced Deprecation Handling', () => {
   beforeEach(() => {
