+++ tests/guidelines/2024-1/item3-display.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 3: Display Attribute Handling', () => {
   beforeEach(() => {
