+++ tests/guidelines/2024-1/item4-alttext-altimg.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 4: Alttext and Altimg Deprecation', () => {
   beforeEach(() => {
