+++ tests/guidelines/2024-1/item5-invisible-operators.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 5: Invisible Operators Updates', () => {
   beforeEach(() => {
