+++ tests/guidelines/2024-1/item6-special-characters.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 6: Special Character Handling', () => {
   beforeEach(() => {
