+++ tests/guidelines/2024-1/item7-chemistry-markup.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 7: Chemistry Markup Support', () => {
   beforeEach(() => {
