+++ tests/guidelines/2024-1/item8-labeled-equations.test.js
@@ -1,4 +1,4 @@
-const { GenerateMath } = require('../../../conversions/text');
+const { GenerateMath } = require('../../../src/conversions/text');
 
 describe('Item 8: Labeled Equations Support', () => {
   describe('Basic Labeled Equations', () => {
